Changing to ints

+++ 190815TDD/Domain/Division.cs
@@ -1,6 +1,6 @@
 namespace TDD.Domain {
     public class Division {
-        public double Dividir(double v1, double v2) {
+        public double Dividir(int v1, int v2) {
             return 1;
         }
     }
